requirements.txt
flask
SQLAlchemy
flask_sqlalchemy
python-dotenv
gunicorn
# Flask-Minify
# Flask-Minify[go] if u aren't using webpack, this could be useful
flask-redis
flask-cors
cryptography
pyjwt
bcrypt
Faker
sentry-sdk[flask]
flask_minify
# newrelic